Remove the builders module, move the logic to models/request_validators
