+++ client/hwlib/src/py_bindings.rs
@@ -18,8 +18,10 @@
  *        Nadzeya Hutsko <nadzeya.hutsko@canonical.com>
  */
 
-use crate::builders::request_builders::{create_certification_status_request, Paths};
-use crate::send_certification_status_request as native_send_certification_status_request;
+use crate::{
+    models::request_validators::{CertificationStatusRequest, Paths},
+    send_certification_status_request as native_send_certification_status_request,
+};
 use once_cell::sync::Lazy;
 use pyo3::{
     exceptions::PyRuntimeError,
@@ -36,7 +38,7 @@ static RT: Lazy<Runtime> = Lazy::new(|| Runtime::new().expect("failed to create 
 /// hardware-api server URL.
 #[pyfunction]
 fn send_certification_request(py: Python, url: String) -> PyResult<PyObject> {
-    let request_body = create_certification_status_request(Paths::default())
+    let request_body = CertificationStatusRequest::new(Paths::default())
         .map_err(|e| PyRuntimeError::new_err(format!("failed to create request: {}", e)))?;
 
     let response =
